refactor client providers structure
